📦 chore(pom.xml): add pom.xml file to project
The pom.xml file is added to the project. This file is used for configuring the project's build and dependencies using Maven. It includes the necessary dependencies for a Spring Boot project, such as spring-boot-starter-web and spring-boot-devtools. The project is set to use Java version 17. The build section includes the spring-boot-maven-plugin for building the project.